m.lab/solution-generator/requirements.txt
streamlit==1.36.0
PyYAML==6.0.1
kaggle==1.6.17
nbformat==5.10.4
openai==1.32.0
pyfiglet==1.0.2
gitpython==3.1.43
pyyaml==6.0.1
pytz==2021.3
boto3==1.34.19
botocore==1.34.19
psutil==5.9.5
pip-tools==7.4.1
requests==2.31.0
redis==5.0.1
docker==7.0.0
tabulate==0.9.0
colorama==0.4.6
pyfiglet==1.0.2
python-dotenv==1.0.1
streamlit-option-menu==0.3.13
streamlit-pills==0.3.0
streamlit-tree-select==0.0.5
streamlit-autorefresh==1.0.1
streamlit-ace==0.1.1
langchain==0.2.5
langchain_community==0.2.5
bs4==0.0.2
langgraph==0.2.3
streamlit-option-menu==0.3.13
tiktoken==0.8.0
chardet==5.2.0
# rich
# ALO requirement template
# -r ./engine/alo_engine/requirements.txt
